+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ### Added
 - License, initial mod setup, mod metadata etc
+- Block highlighting util